% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{code}
\alias{code}
\alias{code,Chunk-method}
\alias{label}
\alias{label,Chunk-method}
\alias{label<-}
\alias{label<-,Chunk-method}
\alias{chunkOptions}
\alias{chunkOptions,Chunk-method}
\alias{chunkOptions<-}
\alias{chunkOptions<-,Chunk-method}
\alias{textAbove}
\alias{textAbove,Chunk-method}
\alias{textAbove<-}
\alias{textAbove<-,Chunk-method}
\alias{textBelow}
\alias{textBelow,Chunk-method}
\alias{textBelow<-}
\alias{textBelow<-,Chunk-method}
\title{Chunk get and set methods}
\usage{
code(x)

\S4method{code}{Chunk}(x)

label(x)

\S4method{label}{Chunk}(x)

label(x) <- value

\S4method{label}{Chunk}(x) <- value

chunkOptions(x)

\S4method{chunkOptions}{Chunk}(x)

chunkOptions(x) <- value

\S4method{chunkOptions}{Chunk}(x) <- value

textAbove(x)

\S4method{textAbove}{Chunk}(x)

textAbove(x) <- value

\S4method{textAbove}{Chunk}(x) <- value

textBelow(x)

\S4method{textBelow}{Chunk}(x)

textBelow(x) <- value

\S4method{textBelow}{Chunk}(x) <- value
}
\arguments{
\item{x}{S4 object of class Chunk}

\item{value}{value to set}
}
\description{
Get and set methods for Chunk S4 class.
}
