% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\docType{class}
\name{Chunk-class}
\alias{Chunk-class}
\title{Chunk S4 class}
\description{
An S4 class for storing code chunk options.
}
\section{Slots}{

\describe{
\item{\code{label}}{the chunk label}

\item{\code{knitr_options}}{a list of knitr chunk options}

\item{\code{code}}{the R code to include in the chunk}

\item{\code{text_above}}{text to include above the code chunk}

\item{\code{text_below}}{text to include below the code chunk}
}}

