% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkify.R
\name{chunkify}
\alias{chunkify}
\title{FUNCTION_TITLE}
\usage{
chunkify()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
