% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkify.R
\name{chunkify}
\alias{chunkify}
\title{Wrap Script in Rmarkdown Chunk}
\usage{
chunkify()
}
\value{
NOTHING
}
\description{
Inputs highlighted text in Rstudio document and wraps it in Rmarkdown Chunk
}
\details{
Use \code{\link{chunky_opts}} to set session parameters for addin
}
