% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nii2fst.R
\name{nii2fst}
\alias{nii2fst}
\title{Convert Nifti Image into FST Fast file}
\usage{
nii2fst(x, fst_file = NULL)
}
\arguments{
\item{x}{an object that \link[RNifti:retrieveNifti]{RNifti::retrieveNifti}}

\item{fst_file}{output filename of FST file}
}
\value{
A file path
}
\description{
Convert Nifti Image into FST Fast file
}
\examples{
arr = array(rnorm(50^3), dim = rep(50, 3))
}
